added collectionservice as a service class to help with submit
